test/Add end-to-end ingredient creation test
Updates UsesDomainService to install the DomainService as a static variable, so classes that subclass UsesDomainService can now use the domain service instance in their constructors. This means, for example, that we don't need to pass a list of units to a class that needs units in the constructor, if the class subclasses UsesDomainService, because it can use the units on the domain service.

Implements flag and nutrient functionality on the ingredient using abstract base classes. This is to pave a route to checking flag and nutrient value validity, by creating a common interface for objects which have flags and have nutrients.

Introduces the UniqueDict class, to make indexing sets of unique items easier.

Updates the concept of Flag to separate out FlagDefinition. Old implementation shared flag instances between ingredients, also sharing flag values, which obviously doesn't work.